Support pulling git tokens from ENV vars
If a token value in the config file matches the format
"ENV.$ENV_VAR_NAME" then the value for the token will be pulled from
the environment variable "$ENV_VAR_NAME". This will make it easier to
store a full config file in git and store the tokens in Openshift
secrets.