Allow sorting reviews by updated date
Sorting reviews by last comment time instead of submitted time allows to keep
track of active and stale reviews. But there are other events that can make
the review count as active, for example, a new patch set in Gerrit.

The updated time property is available in the response from all the review
services, so let's add it to review-rot and allow sorting the reviews
using this value to give a better picture of stale and active reviews.

The --sort argument is added, supporting sort by submitted time, updated
time and last commented.

Sorting by last comment time was already available using the
--sort-comment argument. With the new sort argument, --sort commented is
equivalent, so the former argument is hidden from the help output and
kept for backwards compatibility.

The updated time property is added to the JSON output format, and the
web interface is updated to include this value using the moment
javascript library to show it in relative time.

Test for arguments are updated with the sort option.