Fix JSON formatting.
The way the code previously worked here, if either `i` or `N` were zero, then
we would not render any JSON for the element.

This always occurred on the *first* item in the list, for which `i` is always
zero.  This means our JSON-based reports have been missing their first hidden
entry, since 7fe84e8979.